move some api from skynet.lua to skynet/manager.lua
