+++ examples/abort.lua
@@ -1,3 +1,4 @@
 local skynet = require "skynet"
+require "skynet.manager"	-- import skynet.abort
 
 skynet.abort()
