+++ service/cslave.lua
@@ -1,5 +1,6 @@
 local skynet = require "skynet"
 local socket = require "socket"
+require "skynet.manager"	-- import skynet.launch, ...
 local table = table
 
 local slaves = {}
