+++ service/service_mgr.lua
@@ -1,4 +1,5 @@
 local skynet = require "skynet"
+require "skynet.manager"	-- import skynet.register
 local snax = require "snax"
 
 local cmd = {}
